Upgrade ovs version in Dockerfile to 2.11
Due to the problems met in issue:
https://github.com/ovn-org/ovn-kubernetes/issues/855
We should use ovs 2.11 instead of 2.10 to fix the problem
of ovn-nbctl running in daemon mode.

So the ovs version used in the Dockerfile should be promoted
to 2.11 to build a suitable image for daemonset install.
The DPDK library is also updated 18.11 to satisfy the requirement
of ovs2.11.
The related changes in Dockerfile for rpmArch of aarch64(arm64)
are also given.

Signed-off-by: Trevor Tao <trevor.tao@arm.com>