Upgrade ovs version in Dockerfile.ubuntu to 2.11
Due to the problems met in issue:
https://github.com/ovn-org/ovn-kubernetes/issues/855
We should use ovs 2.11 instead of 2.10 to fix the problem
of using ovn-nbctl in daemon mode(default).

So the ovs version used in the Dockerfile should be promoted
to 2.11 to build a suitable image for daemonset install.
The related changes in Dockerfile.ubuntu are also verified on
aarch64(arm64) platform in addition to amd64(x86_64).

Signed-off-by: Trevor Tao <trevor.tao@arm.com>