cni/linux: ensure only the latest sandbox has external-ids:iface-id set
Kubelet may keep multiple sandboxes running for a given pod if some are
waiting for garbage collection. We need to make sure that only the
latest sandbox has iface-id set to the container's namespace/name or
ovn-controller may get confused about which OVS port to associate with
the Pod's logical switch port.

Signed-off-by: Dan Williams <dcbw@redhat.com>