do not assign cluster-admin role for the `ovn` serviceaccount
the serviceaccount user `ovn` has a cluster-admin role today and can
pretty much do anything in the clsuter. we need to fine grain access
control and pod privileges for this user.

to do this:

1. define a PodSecurityPolicy object that captures minimum required
   security policies to run our deployments and daemonset.
2. define a ClusterRole object that captures all the resources we are
   intersted in and all the actions we need on them. also, this role
   should use the PodSecurityPolicy defined in step 1.
3. bind the above role to `ovn` serviceaccount

Note: this commit adds (1) that provides almost any securityContext to
be set (this can be restricted in future commits)

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>