Bandwidth: Actual egress bandwidth might be much lower than configured.
For egress network bandwidth, we configure ingress_policying_rate,
but, by itself, it might end up allowing less than the rate (e.g.
iperf TCP with a rate of 4Gbps gives < 1Gbps). OVSDB schema has
this to say about ingress_policying_burst:

"
Specifying a larger burst size lets the algorithm be more forgiving,
which is important for protocols like TCP that react severely to
dropped packets. The burst size should be at least the size of the
interface’s MTU. Specifying a value that is numerically at least as
large as 10% of ingress_policing_rate helps TCP come closer to
achieving the full rate
"

Set ingress_policying_burst to 10% of ingress_policying_rate.

Additionally, for a SR-IOV VF we also need to configure its
max_tx_rate using the egress value in Mbps.

Tested by configuring egress-bandwidth for a pod and using iperf.
Tested the VF setting by giving a VF as OVN interface to the pod,
with egress-bandwidth configured.

Signed-off-by: venu iyer  <venugopali@nvidia.com>