allow OVS internal port to be a shared gateway interface
In shared gateway mode, we share one of the interfaces on the node for
L3 gateway functionality (North-South) traffic. The IP address of this
interface on the host will be used to SNAT the pod traffic reaching out
to external world. Any NodePort traffic hitting the node at the shared
interface IP will be steered into OVN topology through OpenFlow rules.

The shared interface itself can be:
1. a physical interface
2. a OVS bridge interface
3. one of the internal ports of the OVS bridge

We support (1) and (2) today, but not ()3. This commit adds support
for (3).

Fixes #883

Signed-off-by: Zhen Wang <zhewang@nvidia.com>
Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>