reduce the timeout from 300 seconds to 60 seconds
i don't think we should wait for 5 mins for ovn-controller to be up
and ready. before the ovnkube-node process starts the ovnkube.sh
script would have waited enough for ovn-controller to be up. all we
are doing here is to check ovn-controller is connected to SB. if it
takes more than a minute to connect, then we have a bigger issue.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>