Support latest OVN
OVN post split from OVS (post 2.12) has changed the paths for the
runtime directory, log dir and db dir.
The default runtime dir is now - /var/run/ovn
The default log dir is now - /var/log/ovn
And the default db dir is - /etc/ovn

In order to support latest OVN version, this patch checks for the
presence of the file - /usr/bin/ovn-appctl. If this utility is
present, it means the deployment has latest OVN with new dir paths
and uses the new dirs. Else, the old openvswitch dir paths are used.

Signed-off-by: Numan Siddique <numans@ovn.org>