running ovn-kube-util shouldn't require binaries that it doesn't use
While executing ovn-kube-util on a host to bring a NIC under an OVS
bridge, it fails with `ovn-nbctl` command not found. The subcommand
only needs `ovs-vsctl` to run, but expects a whole plethora of utilities
to be present. This commit validates only the required CLI binaries to
be present for ovn-kube-util to succeed.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>