Move pod annotation under k8s.ovn.org namespace
The pod network info of IP, MAC, Gateway, and Routes are under
'ovn' annotation. We need to move it under 'k8s.ovn.org' namespace.

The new annotation is called 'pod-networks', and it is going to be a
map of 'network_name' to pod's IP information on that network. For
example: ("default" refers to the first OVN interface to the Pod)

    {
        "default": {
            "gateway_ip": "192.168.2.1",
            "ip_address": "192.168.2.3/24",
            "mac_address": "8a:24:f4:a8:02:04"
        }
    }

The changes assumes that the master is upgraded first. It continues to
write both the old/new annotation names to facilitate yet-to-be
upgraded ovnkube nodes. In the next release of ovn-kubernetes, we can
remove the code that adds `legacy` annotation.

Signed-off-by: Yun Zhou <yunz@nvidia.com>
Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>