Move node annotation under k8s.ovn.org namespace
The node's subnet information is  under 'ovn_host_subnet' annotation.
We need to move it under 'k8s.ovn.org' namespae.

The new annotation is called 'node-subnets', and it is going to be a
map of 'network_name' to node's subnet allocation from that network.
For example: ("default" refers to the first OVN interface to the Pod)
    k8s.ovn.org/node-subnets: {
       "default": "192.168.10.0/24"
    }

The changes assumes that the master is upgraded first. Since this
annotation is used only in the master the old worker nodes are not
going to be affected by this change. As a result, on master upgarde to
this code we are going to remove the old key and add the new key to
the nodes as part of syncNodes() call.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>