persist the mac address associated with the node management port
the MAC address for node's management port is randomly chosen. this
address is then added to node's annotation. the master reads the
address and creates a corresponding logical switch port using this
address.

now when node reboots, the mac address of the management port on the
node changes. this changed address is then reflected on node's
annotation and then in the UpdateFunc callback handler for the node
resource, we update the MAC address of the logical switch port.

this is all unnecessary complexity, so better way is to just
persist the initial MAC for the management port in the interface's
MAC column

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>