gateway_localnet: Use ip6tables if appropriate
Update the gateway_localnet code to conditionally use ip6tables if the
localnet gateway is using IPv6 addresses.  As the code currently
stands, the localnet addresses are hard coded as IPv4.  A future patch
will change that, but this at least puts the check for iptables vs
ip6tables for this code into one spot.

Signed-off-by: Russell Bryant <russell@ovn.org>