no need to add default route to ovn_cluster_router
the logic behind finding the nexthop for default route for
ovn_cluster_router is very involved. it first finds all the L3 gateways
and then sorts it and picks the first one in the sorted list.

we were doing this everytime a node was added. so, in a 400-node cluster
we were doing the same computation and adding the same duplicate route
again (off course with --may-exist flag) 400 times.

so, i thought of caching the result and re-using the cached nexthop.
then it occured to me that we don't even need default route. we already
had source-ip based routes for every node in the cluster. this route
was never used since it will never get matched (as src-ip based route
will always match and the packets from a node will exit out through its
l3 gateway)

so, i have removed the code to add the default route to distributed
router

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>