testing/ovn: don't use k8s.io/utils/exec/testing's FakeExec and print useful output
FakeExec from k8s.io/utils/exec/testing uses panic() in a bunch of places
which horribly confuses the Ginkgo panic recovery code. This leads to panics
from testcase failures that have no Ginkgo output that are extremely hard to
debug.

Instead, since we're already doing most of what FakeExec did in our own
FakeExec wrapper, just do the rest of it. We keep using FakeCmd though
since it's still useful and the panics it has we will never run into.

Additionally, rework the looseCompare logic in our FakeExec to keep the
full list of commands (rather than a command pool) and just mark which ones
have been called, so we can still compare on the total number of commands
run.

Lastly, using new functionality in Gomega v1.8.0 we can pass a function
to matchers that returns a string of additional description, which we
now use to build up the list of expected and executed commands and
indicate which ones were not run, which gets dumped into the Ginkgo
testcase failure message now, rather than the general log output.

Signed-off-by: Dan Williams <dcbw@redhat.com>