set other_config:hwaaddr on br-local before you add br-nexthop
The issue is that br-local was inheriting the lowest MAC from its
port and in this case it was br-nexthop. So, change the order in
which the mac address is fixed for br-local bridge. First, create the
br-local bridge and set other_config:hwaddr with mac_in_use value. Next,
create the br-nexthop interface with fixed mac address.

Fixes: 75f836d91e1e (create br-nexthop OVS internal port with fixed MAC
address)

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>