ovn-controller: Set ovn-openflow-probe-interval
ovn-kubernetes was already setting ovn-remote-probe-interval.  This
patch follows the same pattern for ovn-openflow-probe-interval, and
does it for the same reasons.

The default value for this option is 5 seconds. On a large cluster,
this can cause excessive CPU consumption in ovn-controller.  If it
takes ovn-controller 5 seconds to do a full state computation, then
you'll see ovn-controller end up in effectively a busy loop, because
it isn't able to keep up with this probe interval.

The openflow probe is even less interesting than the OVSDB
remote probe.  At least the ovsdb connection is to something remote.
The openflow connection is always local, so this is unlikely to be a
problem.  We now set it to 3 minutes by default, just in case, instead
of disabling it completely.

Signed-off-by: Russell Bryant <russell@ovn.org>