scale: fixed MAC address for the logical router ports
currently, the MAC address for the logical_router_port on the
distributed_router that connects to node logical_switch is generated
randomly. that by itself is not an issue. however, we depend on this
MAC address on each of the node. each of the ovnkube-node deamon does

   routerMAC=`ovn-nbctl lsp-get-addresses stor-node1`

the above call doesn't scale with large number of nodes, and the above
command times out.

instead of creating a random MAC address, use a fixed MAC address that
is formed using the octets of gateway IP address. we already make an
assumption that the 1st IP of the node subnet is the gateway address,
and therefore we can easily derive at the MAC address on the node
without making a ovn-nbctl call.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>