do normal processing for all inbound packets that are not tracked
Say, we share the interface eth0 in the shared gateway mode and that
the ovs-bridge br-eth0 has an additional interface 'vlan100'

    br-eth0(LOCAL)  vlan100    patch_port_to_br_int
              |        |       |
              |        |       |
            +-+--------+-------+---+
            |       br-eth0        |
            +--------+-------------+
                     |
                     |
                   eth0

In the current code, the inbound traffic to vlan100 is always sent to
LOCAL port and therefore services on vlan100 is not reachable. Fix this
by using NORMAL action.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>