scale: clearInitialNodeNetworkUnavailableCondition() is very expensive
clearInitialNodeNetworkUnavailableCondition() called from Node
UpdateFunc() is very expensive. In our 600-node cluster, the
updateFunc() is called almost every second. We are spending lot of CPU
cycles in that funciton due to DeepCopy and such. call this function if
the node is part of a cloud provider (Node.Spec.ProviderID is set)

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>