exit early if ovn specific pod annotations exists
currently, if the ovn specific pod annotation exists, then we read
pod interface info from the annotation and re-create the corresponding
logical switch port. this is all good, however we continue further and
get the pod interface info from OVN logical topology and add the
annotation back. this is not required. so, exit early

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>