skip adding default gateway for OVN interface in certain cases
It’s generally expected that the the cluster-wide default network
attachment sets a default route. This default route may be overridden
using the “default-route” key in the Network Attachment Selection
Annotation (see 4.1.2.1.9) on another attachments. In that case, we
should skip adding default route on the OVN interface and instead just
add cluster subnet routes for the OVN interface.

Consider the set of annotations below on a pod:

  annotations:
          v1.multus-cni.io/default-network: ovn-kubernetes
          k8s.v1.cni.cncf.io/networks: |
            [
              {
                "name":"sriov-storage-net",
                "default-route": ["10.1.32.2"]
              }
            ]

The first interface to the pod is provided by ovn-kubernetes and the
second interface is through SRIOV CNI. The default route for that
pod should be through the second interface. So, the routes should be
something like this:

$ ip ro
default via 10.1.32.2 dev net1
10.1.32.0/19 dev net1 proto kernel scope link src 10.1.32.142
192.168.0.0/24 dev eth0 proto kernel scope link src 192.168.0.5
192.168.0.0/16 via 192.168.0.1 dev eth0

192.168.0.0/16 is the OVN Kubernetes Cluster Subnet for which the POD
will have a route through OVN's interface (aka eth0)

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>