ovn: read pod addresses more quickly
Previously when adding the pod's logical switch port with the --wait=sb option
the --wait=sb waited long enough for ovn-northd to perform address assignment
for the pod, so the address was immediately available after the lsp-add.

But without the --wait=sb we try to read the address immediately, fail
because ovn-northd hasn't quite gotten around to assigning it yet, and
then wait a full second before we try again.

Instead of waiting a full second, try reading the pod's dynamic addresses
a couple of times quickly, then fall back to per-second polling.

Signed-off-by: Dan Williams <dcbw@redhat.com>