Allow IPv6 Node addresses for geneve tunnels
GetNodeIP is used to determine an IP address for the Node to use as
the address other Nodes will use to create Geneve tunnels to this
Node.  It previously filtered out IPv4 addresses, but that breaks an
environment where the Nodes only have IPv6 addresses.

The code will now adjust between expecting an IPv4 or IPv6 address
based on which mode it is in.

I did not hit this issue when testing ovn-kubernetes with IPv6 on AWS
or Azure before because in those environments, the nodes have both
IPv4 and IPv6 addresses.  I hit this issue when running ovn-kubernetes
in an IPv6-only bare metal environment.

Signed-off-by: Russell Bryant <russell@ovn.org>