Restrict address type requirements for tunneling
This is a follow-up based on discussion in PR #1013.

Relax the requirements on address type for tunnels and use whatever IP
we can find that should work.  If this behavior doesn't work in some
environment, the encap IP can be configured explicitly instead by
setting the encap-ip configuration option.  This should be safe
default behavior, because it will only use an IP that can be
determined via DNS for the Node.

Signed-off-by: Russell Bryant <russell@ovn.org>