a way to specify the encap-ip for a k8s worker node
it could be that the IP that the hostname resolves to is not the
intended ip for Geneve VTEP. provide a way for ovnkube.sh to get to
know about the actual encap IP to be used on the node.

the assumption here is that some entity sets up the
external_ids:ovn-encap-ip with the right vtep ip and ovnkube.sh scraps
that info and passes it to onvkube through --encap-ip option.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>