CARRY: fix race updating exclude IPs on node switch
When nodes are added to the cluster, it may be that our pod add
handler is waiting for the node switch to show up. It might win
the race and create the pod between our node switch being
created in master.go and the exclude IPs being updated immediately
after. Fix this by reverting to setting the exclude IPs during
logical switch creation.

This is essentially a revert of part of 80d6dff11

Fixes: 80d6dff11983a4358693eef06467b1390e96e880