create address_sets only when required for network policy rules
when we create a network policy for ingress (or egress) that contains
only ipBlock fields and nothing else, we end up creating address_sets.
for example: say we have 3 ingress rules and each rule only captures the
ipBlock and nothing else, we end up creating 3 address_sets (since 3
ingress rules) and they don't have any IP addresses. this also results
in creation of 6 ACL - 3 ACLs for the ipBlock CIDR and 3 ACLs with
empty address_set.

the fix in this commit reduces lot of unnecessary ACLs

Fixes: #718
Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>