network policy's logicalport cache gets out-of-sync on pod deletion
when a pod is deleted, its logical_switch_port is deleted and so is the
corresponding entry from logicalPortUUIDCache. In the
handleLocalPodSelectorAddFunc, we fail to get the UUID for the just
deleted pod's logical switch port and we return early from the function
without cleaning up various caches.

Now, say the same pod is added back. All these caches will have stale
data and we fail to add the newly added logical_switch_port to the
port_group.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>