stop delivering ovn-debug.sh script
this script has lot of things that are not right and it doesn' work, so
stop delivering this file. for instance, running 'stop' command on a
daemon from that container just kills the container and restarts the
container. if you can't stop, then you can't start.

lot of output that this script provides can be obtained from the host by
inspecting /var/run/openvswitch or /var/log/openvswitch folders.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>