Enable userspace mode for KIND deployments
This patch enables using netdev mode for OVS. This allows multiple
docker containers hosting different OVS instances to function without
the potential collisions that would occur while sharing the same kernel
data path.

Note, netdev without DPDK is considered "unsupported" officially, and is
something we only want to use for KIND deployments. Therefore the config
option to enable it is hidden, using an environment variable that is not
exposed in the ovn-kubernetes config.

Netdev mode does not support TX checksum offload, therefore it needs to
be disabled on pod veth interfaces as well as veth interfaces attached
from OVS to the host. See:
https://github.com/vmware-tanzu/antrea/issues/14

Co-Authored-by: Andrew Sun <asun@redhat.com>

Signed-off-by: Tim Rozet <trozet@redhat.com>