ipv6: Make IPAddrToHWAddr() aware of IPv6
I had a problem in an IPv6 cluster where certain traffic was getting
dropped.  A closer look revealed that duplicate MACs had been
configured in the OVN virtual network topology.  In particular, the
join bridge between the gateway router and distributed cluster router
had the same MAC on both of its ports.

The root cause was this function.  It assumed IPv4.

While working on a fix for this and writing test cases, I discovered
that the function would sometimes also break for IPv4.  It assumed
that the IPv4 address was always in 4-byte form.  Sometimes the IP
type actually has 16 bytes, even when holding an IPv4 address.  The
first fix is to ensure that we're always using the 4-byte
representation when working with the IPv4 address.

The second fix is to make this function aware of IPv6.  We have more
bytes in the address than we have bytes in a MAC, so it's not
technically possible to always guarantee this results in a unique mac.
However, the subnets we have hard coded internally in OVN are of a
pretty simple format like fd98::N, so using the first two and last two
bytes seems good enough.

Really it's better to just generate a random unique MAC, but this
patch attempts a compromise by covering some cases without
re-introducing the performance concern that was addressed when this
code was first introduced by commit
b453f7398653b9e663866cd337c0d01dbfd80758.

Signed-off-by: Russell Bryant <russell@ovn.org>