deadlock in ovnkube node while waiting on ready functions to complete
we have two ready functions that check for management port and gateway
readiness. if both of them timeout, then they try to write to an
unbuffered `errors` channel. one of the function will succeed while
the other function gets blocked in

  w.errors <- err

the main goroutine gets blocked waiting for both the ready functions to
complete, however one of them cannot complete as it is blocked.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>