readiness-probe for OVN NB and SB Raft ovsdb-server
ovsdb-server 2.9 or later introduced a database called _Server. this
database exists in every instance of ovsdb-server process. this database
has a table called Database and the rows in that table captures the
status and configuration of every database ovsdb-server handles
(including _Server database)

for example, in the case of NB ovsdb-server there will be two databases
OVN_Northbound and _Server. the _Server`Database table will have two
rows. one of the columns in the row is called Connected and it is of
boolean type and for cluster database it reprsents whether the server
is in contact with a majority of its cluster.

checking on `connected` to be true is a good readiness probe for both
NB and SB container.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>