Adds SCTP protocol support to services
This patch will detect if the current OVN version supports SCTP and then
enable it for K8S services if so. If no support is detected and a user
creates an SCTP service in k8s, a warning event will be recorded for the
resource in k8s.

Also cleans up and consolidates a bunch of duplicate logic for each
protocol. Additionally, removes using self-defined constants for
protocols and uses kapi versions for services.

Closes: #1120
Closes: #1124

Signed-off-by: Tim Rozet <trozet@redhat.com>