Remove some unused pod routing code
The CNI code had code to check if there was a default route specified
in the "routes" field of the pod annotation and ignore the
"gateway_ip" field in that case, but we never actually used that
functionality. Remove the code, and sanity-check the routes when
marshalling/unmarshalling the pod annotation.

Signed-off-by: Dan Winship <danw@redhat.com>