Fix nbctld log-file
OVN_LOG_NBCTLD env var was not used in ovnkube-master deployment,
so interally in ovnkube.sh it was not mapping to anything, and hence
it was setting to right log file name. Once we injected this env var
through patch (35f27054) it was mapped to log-file and not log-level
and configured --log-file incorrectly.

Signed-off-by: Anil Vishnoi <vishnoianil@gmail.com>