config: fix test cases
We were passing "unaligned" CIDR values (eg, "10.129.0.0/14", which is
actually an address inside the "10.128.0.0/14" block), and thus we
weren't even actually overriding the values, and the test only worked
at all because the test case used the same code to parse the strings
as the actual config code does.

Anyway, fix it to use valid strings.

Signed-off-by: Dan Winship <danw@redhat.com>