identify default flows with 0xdeffl05 cookie and add healthcheck for it
in the shared gateway mode, we share the host's k8s node ip for SNATing.
we add default set of 'five' flows to steer the nodeport traffic into
the OVN logical topology and to also connection track the l3 gateway
north-south traffic

these flows are very critical and if they are missing for whatever
reason, then that node is not functional. so, we need to add check for
it and take corrective action (for now just kill the ovnkube daemon
so that it restarts with right flows).

Signed-off-by: Yun Zhou <yunz@nvidia.com>
Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>