util: fix handling of ovn-nbctl daemon mode paths
ovn-nbctl daemon mode by default puts its pidfile and socket
into /var/run/openvswitch which is likely an oversight in the
OVS/OVN split work debuting in OVN 2.12. The code was not looking
in /var/run/openvswitch if ovn-appctl was present, which was supposed
to be the magic way of figuring out whether to use /var/run/openvswitch
or /var/run/ovn.

Instead, first look for an existing OVN_NB_DAEMON environment variable,
which may have been set already per 'man ovn-nbctl' daemon mode
documentation, and return that verbatim to ensure the child ovn-nbctl
process uses it.

Otherwise, look for the pidfile and corresponding socket in both
directories (ovn first, then openvswitch) and pick the first one
that has both.

Signed-off-by: Dan Williams <dcbw@redhat.com>