dual-stack support in KIND
This PR enables dual-stack in a KIND deployment using OVN-Kubernetes. Summary
of changes:
* Added contrib/dualstack.md to explain how to run with dual-stack in a KIND
deployment.
* Update contrib/kind.sh script to enable dual-stack in KIND.
* Added contrib/dualstack.sh script to manage the more complex sed commands
that are used to update the kind.yaml file. Called from kind.sh.
* Moved contrib/kind.yaml to contrib/kind-noha.yaml. In the current
implementation the kind.yaml was edited in place and at a minimum had a change
to 'apiServerAddress: 11.12.13.1' occured when kind.sh was run. kind-noha.yaml
and kind-ha.yaml are now copied over to kind.yaml (which is no longer checked
in).

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>