Update to urfave/cli/v2
This commit updates the dependency on urfave/cli to v2.
v2 is the next mainline version.
One particularly useful feature that I'm interested in the `RunContext`
method that allows for passing a context to the CLI commands such that
cancellation and timeouts can be handled.
This is useful to implement proper signal handling for the controller.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>