remove OVN specific podsecuritypolicy resource
the podsecuritypolicy is a cluster-level resource and we were defining
it in ovn-setup.yaml. when PodSecurityPolicy admission controller was
enabled, all the non-ovn-k8s pods were getting tagged with this policy
and that definitely was not the intent when it was initially added

the workflow would be that the cluster-admin would define a policy for
the whole cluster and add 'use' verb for the polices definied for the
'ovn' service account. in short, it should come from the cluster
deployer and not from us

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>