remove redundant log message that fills up the onvkube-master log file
if we fail to use the ovn-nbctl daemon mode for whatever reason, then
we throw a warning message. if this warning message is missing, then
the command was executed using the ovn-nbctl in daemon mode. so, no need
to log the info that we are using ovn-nbctl daemon mode. this can be
inferred from missing warning message

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>