fix CNIRequestMetrics that tracks CNI requests (ADD, DEL) duration
the flow of CNIAdd/CNIDel requests from either kubelet or multus is as
shown below

+--------+      CNIShim                                  CNIServer
|Kubelet/|   +-----------+  /var/run/ovn-kubernetes/  +--------------+
| Multus |   | ovn-k8s-  |   cni/ovn-cni-server.sock  | ovnkube-node |
| exec's |-->|cni-overlay|--------------------------->|  httpServer  |
+--------+   +-----------+          AF_UNIX           +--------------+
   (a)            (b)               Socket                    (c)

if OVS daemons are running inside the container, then all the operations
-- namespace creation, VETH creation, plugging the interface into the
POD, and creation and additon of OVS internal port occurs inside the
container (c). This requires the container to be `privileged`.

In the other mode, aka unprivileged mode, everything happens in (b). We
just get the pod interface information from (c) and all the plumbing
happens in (b).

The current metric computation was happening in (c) and was showing
incorrect metrics for the unprivileged case.

This commit computes the `elapsedTime` for a given command in (b) and
finally posts the value to http://dummy/metrics endpoint for (c) to
publish the prometheus metric.

This captures the true end-to-end delay in creating the interface or
deleting the interface.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>